pyproject.toml
[tool.poetry]
name = "repl_python3_beta_clooxx"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
dnspython = "^2.1.0"
Flask = "^2.0.2"
pymongo = "^3.12.0"
Werkzeug = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
